requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
google-generativeai==0.8.2
python-dotenv==1.0.1
scikit-learn>=1.5.0
PyPDF2>=3.0.0
python-docx>=1.1.0
itsdangerous>=2.1.2
psycopg2-binary>=2.9.0
